Move language-specific code out of the util/__init__.py and into its own module.
