+++ tests/util/test_string_helpers.py
