+++ tests/util/test_web_publication_manifest.py
